refactor(AdminRepository): move related methods to DriverRepository and TripRepository
