+++ src/TriQue.App/Data/Repositories/TripRepository.cs
@@ -1,6 +1,4 @@
 ﻿using Microsoft.Data.Sqlite;
-using System;
-using System.Collections.Generic;
 using System.Data;
 using TriQue.Enums;
 using TriQue.Helpers;
@@ -17,6 +15,7 @@ namespace TriQue.Data.Repositories
             _dbHelper = new DatabaseHelper();
         }
 
+        #region Trip Retrieval
         public List<Trip> GetByDriverID(int driverID)
         {
             var trips = new List<Trip>();
@@ -50,6 +49,63 @@ namespace TriQue.Data.Repositories
 
             return trips;
         }
+
+        public int? GetActiveTripID(int driverID)
+        {
+            string query = @"
+                SELECT TripID FROM Trip
+                WHERE DriverID = @driverID
+                  AND StatusID = 2
+                  AND EndTime IS NULL
+                ORDER BY TripID DESC
+                LIMIT 1";
+
+            var result = _dbHelper.ExecuteScalar(
+                query,
+                new SqliteParameter("@driverID", driverID)
+            );
+
+            return result == null ? null : Convert.ToInt32(result);
+        }
+
+        #endregion
+
+        #region Trip Actions
+
+        public void StartTrip(int driverID, int routeID)
+        {
+            string query = @"
+                INSERT INTO Trip (DriverID, RouteID, StatusID, ActualEarnings, StartTime)
+                    VALUES (@driverID, @routeID, 2, 0, @startTime)";
+
+            _dbHelper.ExecuteNonQuery(
+                query,
+                new SqliteParameter("@driverID", driverID),
+                new SqliteParameter("@routeID", routeID),
+                new SqliteParameter("@startTime", DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss"))
+            );
+        }
+
+        public void EndTrip(int tripID, double fare)
+        {
+            string query = @"
+                UPDATE Trip
+                SET EndTime = @endTime,
+                    StatusID = 3,
+                    ActualEarnings = @fare
+                WHERE TripID = @tripID";
+
+            _dbHelper.ExecuteNonQuery(
+                query,
+                new SqliteParameter("@endTime", DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss")),
+                new SqliteParameter("@fare", fare),
+                new SqliteParameter("@tripID", tripID)
+            );
+        }
+
+        #endregion
+
+        #region Driver Stats
 
         public double GetEarningsProgress(int driverID)
         {
@@ -121,55 +177,21 @@ namespace TriQue.Data.Repositories
             return (fastest, slowest);
         }
 
-        public int? GetActiveTripID(int driverID)
+        #endregion
+
+        #region Reports
+
+        public int GetTotalTrips()
         {
-            string query = @"
-                SELECT TripID FROM Trip
-                WHERE DriverID = @driverID
-                  AND StatusID = 2
-                  AND EndTime IS NULL
-                ORDER BY TripID DESC
-                LIMIT 1";
+            string query = @"SELECT COUNT(*) FROM Trip";
 
-            var result = _dbHelper.ExecuteScalar(
-                query,
-                new SqliteParameter("@driverID", driverID)
-            );
+            using var conn = _dbHelper.GetConnection();
+            conn.Open();
 
-            return result == null ? null : Convert.ToInt32(result);
+            using var cmd = new SqliteCommand(query, conn);
+
+            return Convert.ToInt32(cmd.ExecuteScalar());
         }
-
-        public void StartTrip(int driverID, int routeID)
-        {
-            string query = @"
-                INSERT INTO Trip (DriverID, RouteID, StatusID, ActualEarnings, StartTime)
-                    VALUES (@driverID, @routeID, 2, 0, @startTime)";
-
-            _dbHelper.ExecuteNonQuery(
-                query,
-                new SqliteParameter("@driverID", driverID),
-                new SqliteParameter("@routeID", routeID),
-                new SqliteParameter("@startTime", DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss"))
-            );
-        }
-
-        public void EndTrip(int tripID, double fare)
-        {
-            string query = @"
-                UPDATE Trip
-                SET EndTime = @endTime,
-                    StatusID = 3,
-                    ActualEarnings = @fare
-                WHERE TripID = @tripID";
-
-                    _dbHelper.ExecuteNonQuery(
-                        query,
-                        new SqliteParameter("@endTime", DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss")),
-                        new SqliteParameter("@fare", fare),
-                        new SqliteParameter("@tripID", tripID)
-                    );
-                }
-
         public DataTable GetTripHistory(int driverID)
         {
             string query = @"
@@ -241,17 +263,7 @@ namespace TriQue.Data.Repositories
             dt.Load(cmd.ExecuteReader());
             return dt;
         }
-        public int GetTotalTrips()
-        {
-            string query = @"SELECT COUNT(*) FROM Trip";
 
-            using var conn = _dbHelper.GetConnection();
-            conn.Open();
-
-            using var cmd = new SqliteCommand(query, conn);
-
-            return Convert.ToInt32(cmd.ExecuteScalar());
-        }
 
         // panels in generating reports
         public (int totalTrips, double totalEarnings, string mostActive, string leastActive, double fastest, double slowest)
@@ -360,5 +372,77 @@ namespace TriQue.Data.Repositories
 
             return (totalTrips, totalEarnings, mostActive, leastActive, fastest, slowest);
         }
+
+        #endregion
+
+        #region Route Stats
+
+        public string GetTotalTripsTodayRoute()
+        {
+            string query = @"
+                SELECT r.RouteName, COUNT(t.TripID) as Total
+                FROM Trip t
+                JOIN Route r ON t.RouteID = r.RouteID
+                WHERE DATE(t.StartTime) = DATE('now', 'localtime')
+                GROUP BY t.RouteID
+                ORDER BY Total DESC
+                LIMIT 1
+            ";
+
+            using var conn = _dbHelper.GetConnection();
+            conn.Open();
+
+            using var cmd = new SqliteCommand(query, conn);
+            using var reader = cmd.ExecuteReader();
+
+            return reader.Read() ? reader.GetString(0) : "No Trips Yet";
+        }
+
+        public (string routeName, int count) GetHighestTripsRoute()
+        {
+            string query = @"
+                SELECT r.RouteName, COUNT(t.TripID) as Total
+                FROM Trip t
+                JOIN Route r ON t.RouteID = r.RouteID
+                GROUP BY t.RouteID
+                ORDER BY Total DESC
+                LIMIT 1
+            ";
+
+            using var conn = _dbHelper.GetConnection();
+            conn.Open();
+
+            using var cmd = new SqliteCommand(query, conn);
+            using var reader = cmd.ExecuteReader();
+
+            return reader.Read()
+                ? (reader.GetString(0), reader.GetInt32(1))
+                : ("No Data", 0);
+        }
+
+        public (string routeName, int count) GetLowestTripsRoute()
+        {
+            string query = @"
+                SELECT r.RouteName, COUNT(t.TripID) as Total
+                FROM Trip t
+                JOIN Route r ON t.RouteID = r.RouteID
+                GROUP BY t.RouteID
+                ORDER BY Total ASC
+                LIMIT 1
+            ";
+
+            using var conn = _dbHelper.GetConnection();
+            conn.Open();
+
+            using var cmd = new SqliteCommand(query, conn);
+            using var reader = cmd.ExecuteReader();
+
+            return reader.Read()
+                ? (reader.GetString(0), reader.GetInt32(1))
+                : ("No Data", 0);
+        }
+
+        #endregion
+
     }
 }
